Add Battery Support (Enphase) Upgrade to Spring Boot 3.5.14 Upgrade Jars Upgrade H2 database to 2.3 and add scripts to upgrade in Dockerfile Support latest Enphase Firmware with HTTPS everywhere

+++ src/main/java/com/hz/services/PrometheusService.java
@@ -29,7 +29,7 @@ public class PrometheusService {
 				.stream()
 				.filter(metric -> metric.getName().equalsIgnoreCase(name))
 				.findFirst()
-				.map(metric -> BigDecimal.valueOf(metric.getValue()))
+				.map(metric -> metric.getValue())
 				.orElse(BigDecimal.ZERO).doubleValue();
 	}
